requirements.txt
chromadb>=0.5.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
pillow>=10.0.0
pypdf>=4.0.0
pytesseract>=0.3.10
sentence-transformers>=3.0.0
streamlit>=1.36.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.0.0
odfpy>=1.4.1
pytest>=8.0.0
torchvision>=0.20.0
